requirements.txt
httpx>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
